Further reduce the code to the absolute minimum necessary.
This plus Hide Tab Bar On One Tab is now a working solution for me
though I'd like to roll Hide Tab Bar into this one so I don't need it.

This is tricky: I have learned that that extension is a XUL extension,
and XUL is being deprecated, which means I need to port its code.
Also, the source isn't online, so I had to unpackage it from the installed file.
Also the source is more complicated than necessary: it runs some javascript in an updateVisibility handler
when all you need to do (maybe this is a newish feature) is use CSS on the Chrome:
```
@-moz-document url-prefix('chrome://') {
.tabbrowser-tabs { display: none; }
}
```
(though this causes glitchiness in Firefox if you can still make extra tabs;
 but the point stands, and I'm sure I can replicate the trick using the new
 https://developer.mozilla.org/en-US/Add-ons/SDK)