Squash the page-blanking bug.
It turns out that some sequence of events deep in the bowels of Firefox will
cause it to expect to be able to read the properties of the .tabbrowser-tabs
XUL element (that is the GUI of the list of tabs) and not having it will cause
the rendering to crash hard and the page to stop responding.

The solution is to revert to exactly what:
now that I know more of my way around this code, I could understand how to port
this XUL overlay
```
        <binding id="hidetabbar-tabbrowser-tabs" extends="chrome://browser/content/tabbrowser.xml#tabbrowser-tabs">
                <implementation implements="nsIDOMEventListener">
                        <method name="updateVisibility">
                                <body><![CDATA[
                                        this.visible = this.tabbrowser.visibleTabs.length > 1;

                                        if (document.getElementById("history-panelmenu"))
                                                document.getElementById("history-panelmenu").open = false;
                                ]]></body>
                        </method>
```

to this Javascript:
```
        let tabbar = getTabContainer(viewFor(window));
        tabbar.updateVisibility = function() { this.visible = false; }
        tabbar.updateVisibility();
```
(both of them replace the updateVisibility method of the tab container)

Firefox never expected me to be tampering with these pieces of its internals,
so maybe it's fair for it to crash, though the /way/ it crashes is reallly bad.